업데이트 후 윈도우 생성 오류

+++ WinAPI/GameEngineBase/GameEngineMath.h
@@ -40,12 +40,12 @@ public:
 		return Y * 0.5f;
 	}
 
-	inline float ihX() const
+	inline int ihX() const
 	{
 		return static_cast<int>(hX());
 	}
 
-	inline float ihY() const
+	inline int ihY() const
 	{
 		return static_cast<int>(hY());
 	}
